split bench code into setup.rs and bench.rs
